♻️ Refactores parser
For #4
