+++ src/Entity/Item.php
@@ -2,7 +2,7 @@
 
 namespace Shikiryu\SRSS\Entity;
 
-class Item
+class Item implements SRSSElement
 {
     public ?string $title;
     public ?string $link;
@@ -15,5 +15,20 @@ class Item
     public ?string $pubDate;
     public ?string $source;
 
+    /**
+     * @var \Shikiryu\SRSS\Entity\Media\Content[]
+     */
+    public array $medias = [];
+
     public array $required = ['description'];
+
+    public function isValid(): bool
+    {
+        return count(array_filter($this->required, fn($field) => !empty($this->{$field}))) === 0;
+    }
+
+    public function toArray(): array
+    {
+        return get_object_vars($this);
+    }
 }
