+++ src/Entity/SRSSElement.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Shikiryu\SRSS\Entity;
+
+interface SRSSElement
+{
+    /**
+     * @return bool
+     */
+    public function isValid(): bool;
+
+    /**
+     * @return array
+     */
+    public function toArray(): array;
+}
